+++ src/components/app/ProjectPreviewSet.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import type { LocaleTextAccessor } from '@locale/Locales';
     import { type Snippet } from 'svelte';
-    import { locales } from '@db/Database';
+    import { disconnected, locales } from '@db/Database';
     import type Project from '@db/projects/Project';
     import Button from '@components/widgets/Button.svelte';
     import ConfirmButton from '@components/widgets/ConfirmButton.svelte';
@@ -89,6 +89,7 @@
                         action={() =>
                             removeMeta ? removeMeta.action() : undefined}
                         icon={removeMeta.label}
+                        enabled={!$disconnected}
                         background
                     ></ConfirmButton>{/if}</div
             >{@render children?.()}</ProjectPreview